// src/pages/mayor/type.ts
export type MayorPosition = {
  movimiento_interno_id: string;
  movimiento_interno_code: string;
  movimiento_interno_image: string;
  position: number;
};

export type MayorVotes = {
  movimiento_interno_id: string;
  movimiento_interno_code: string;
  movimiento_interno_image: string;
  votes: number;
};

export type MayorResumen = {
  movimiento_interno_id: string;
  movimiento_interno_code: string;
  movimiento_interno_image: string;
  positions: number;
};